refactor(homepage): MemberTabType, ApprovalTabType을 상수 파일로 이동 (#212)
