--- apps/homepage/src/schemas/admin/admin.type.ts
@@ -1,2 +1,0 @@
-export type MemberTabType = 'ALL' | 'ACTIVE';
-export type ApprovalTabType = 'REQUESTED' | 'REJECTED';
